don't let package paths be 404s
These now provide the right go-get metadata, and redirect browsers to
pkg.go.dev.

The index pages are practically the same. They're now in directories,
for consistency.

We use a generator, since GitHub Pages is static content. This also
allows caching and the use of CDNs.

We use bash here since the script is reasonably simple. Go would be a
bit overkill given that I probably won't have to test this or maintain
it for other people.

Fixes #2.